chore: move the collector to integ test dir
